split docker_api_swarm_test.go into multiple files
